4.0.0 BETA 3 [+] Better and more efficient entity check [+] PET needs recoding - perhaps from scratch [-] Jumping needs RECODE 4.0.0 BETA 4 [+] PET Leveling properly added in [+] PET Gears Passive / Guard work as needed ShockToday at 5:15 PM 4.0.0 BETA 5 [+] Improved performance, less lag, less usage [+] Portals should work

+++ NettyBaseReloaded/Game/objects/world/characters/Level.cs
@@ -18,7 +18,7 @@ namespace NettyBaseReloaded.Game.objects.world.characters
             return levelsForCurrentExp;
         }
 
-        public Level DeterminePetLvl(int get)
+        public Level DeterminePetLvl(double get)
         {
             var levelsForCurrentExp = PetLevels.FirstOrDefault(x => x.Value.Experience >= get).Value;
             return levelsForCurrentExp;
